+++ packages/hypergraph/src/mapping/index.ts
@@ -1,0 +1,2 @@
+export * from './Mapping.js';
+export * from './Utils.js';