+++ packages/hypergraph-react/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @graphprotocol/hypergraph-react
 
+## 1.0.0
+### Minor Changes
+
+- 8622688: Move @graphprotocol/typesync Mapping and Utils into @graphprotocol/hypergraph package and export from there. Update @graphprotocol/hypergraph-react to use mapping from @graphprotocol/hypergraph.
+  
+  
+  ## Changes needed
+  
+  Any use of `@graphprotocol/typesync` should use the exported mapping and utils from `@graphprotocol/hypergraph` instead.
+  
+  ### Example
+  
+  ```ts
+  // before
+  import type { Mapping } from '@graphprotocol/typesync/Mapping'
+  
+  // after
+  import type { Mapping } from '@graphprotocol/hypergraph/mapping'
+  ```
+
+### Patch Changes
+
+- Updated dependencies [8622688]
+  - @graphprotocol/hypergraph@0.2.0
+
 ## 0.1.0
 ### Patch Changes